comment out username completino code with macros since unimplemented, fix -Wall warnings on Irix
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@290 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ search.c
@@ -504,11 +504,11 @@ int do_replace_loop(char *prevanswer, filestruct *begin, int *beginx,
 	if (wholewords)
 	{
 	    /* start of line or previous character not a letter */
-	    if ((current_x == 0) || (!isalpha(fileptr->data[current_x-1])))
+	    if ((current_x == 0) || (!isalpha((int) fileptr->data[current_x-1])))
 	    {
 		/* end of line or next character not a letter */
 		if (((current_x + strlen(prevanswer)) == strlen(fileptr->data))
-			|| (!isalpha(fileptr->data[current_x + strlen(prevanswer)])))
+			|| (!isalpha((int) fileptr->data[current_x + strlen(prevanswer)])))
 		    ;
 		else
 		    continue;
